requirements.txt
python-docx>=0.8.11
pillow>=9.0.0
chardet>=4.0.0
pathspec>=0.9.0
sv-ttk>=2.5.0
pyinstaller>=5.6.0
scikit-learn>=1.0.2
numpy>=1.22.0
jieba>=0.42.1
simhash>=2.1.2 